Introduce `WasmInputStream` for reading WASM files
* Parse all sections streamwise
* Do not keep any portion of the file in memory during parsing other than what is needed to parse the current element
* Ensure that all sections are parsed safely even if there is extraneous data

+++ wasm/src/main/java/com/dylibso/chicory/wasm/types/StartSection.java
@@ -1,17 +1,17 @@
 package com.dylibso.chicory.wasm.types;
 
 public class StartSection extends Section {
-    private long startIndex;
+    private int startIndex;
 
     public StartSection() {
         super(SectionId.START);
     }
 
-    public long startIndex() {
+    public int startIndex() {
         return startIndex;
     }
 
-    public void setStartIndex(long startIndex) {
+    public void setStartIndex(int startIndex) {
         this.startIndex = startIndex;
     }
 }
